refactor: isolate searchbar into component
